Extracting Repository object
